Output errors as proper GitHub Actions error messages
As defined in https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#setting-an-error-message

To facilitate this, a helper shell function named 'fail' is defined in a
file on the runner. This file is sourced by all the steps that need the
'fail' function.

This same file may also be used for other common definitions of a job.