Replace env vars with outputs
Replace global env vars to pass information from one step to another
with outputs. The reason is that env vars set in the action will also be
visible in subsequent steps of the calling workflow job (where they
might collide with other env vars that are set in the job).

With outputs, this can't happen. In fact, any outputs of a step inside
the action are not visible outside of the action.