Add 'github-token' input
If this input is set to a valid OAuth2 GitHub API access token (such as
a personal access token or the GITHUB_TOKEN secret of GitHub Actions),
then this token is used to authenticate all GitHub API requests. If this
input is not set, then all GitHub API requests are carried out in
unauthenticated form.

Setting this input may make sense when using a macOS runner because
apparently, all macOS runners share the same IP address (see
https://github.com/actions/virtual-environments/issues/602#issuecomment-602472951).
That means, it's very common to hit the rate limit when making
(unauthenticated) GitHub API requests from a macOS runner.

However, the rate limit for authenticated GitHub API requests is much
higher than for unauthenticated ones. So, when using this input to make
authenticated GitHub API requests on macOS runners, the risk for rate
limit errors is greatly reduced.